Ignoring TTL Deletes from DynamDB

+++ src/main/java/com/jlhood/ddbcopier/DynamoDBCopier.java
@@ -4,6 +4,7 @@ import java.util.Set;
 import java.util.function.Consumer;
 
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
+import com.amazonaws.services.dynamodbv2.model.Identity;
 import com.amazonaws.services.dynamodbv2.model.PutItemRequest;
 import com.amazonaws.services.lambda.runtime.events.DynamodbEvent;
 
@@ -21,6 +22,7 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 public class DynamoDBCopier implements Consumer<DynamodbEvent> {
     public static final String DELETE_EVENT_NAME = "REMOVE";
+    public static final Identity TTL_IDENTITY = new Identity().withPrincipalId("dynamodb.amazonaws.com").withType("Service");
 
     private static final Set<String> ALLOWED_STREAM_VIEW_TYPES = ImmutableSet.of("NEW_IMAGE", "NEW_AND_OLD_IMAGES");
 
@@ -30,8 +32,7 @@ public class DynamoDBCopier implements Consumer<DynamodbEvent> {
     @Override
     public void accept(final DynamodbEvent dynamodbEvent) {
         log.info("Copying {} records", dynamodbEvent.getRecords().size());
-        dynamodbEvent.getRecords().stream()
-                .forEach(this::processRecord);
+        dynamodbEvent.getRecords().forEach(this::processRecord);
     }
 
     private void processRecord(final DynamodbEvent.DynamodbStreamRecord record) {
@@ -39,6 +40,10 @@ public class DynamoDBCopier implements Consumer<DynamodbEvent> {
                 "ddb-copier requires source table stream to be configured with one of the following StreamViewTypes: " + ALLOWED_STREAM_VIEW_TYPES);
 
         if (isDelete(record)) {
+            if (isTTLDelete(record)) {
+                log.info("Ignoring TTL delete on item: {}", record.getDynamodb().getKeys());
+                return;
+            }
             deleteRecord(record);
         } else {
             putRecord(record);
@@ -47,6 +52,10 @@ public class DynamoDBCopier implements Consumer<DynamodbEvent> {
 
     private boolean isDelete(final DynamodbEvent.DynamodbStreamRecord record) {
         return DELETE_EVENT_NAME.equals(record.getEventName());
+    }
+
+    private boolean isTTLDelete(final DynamodbEvent.DynamodbStreamRecord record) {
+        return TTL_IDENTITY.equals(record.getUserIdentity());
     }
 
     private void deleteRecord(final DynamodbEvent.DynamodbStreamRecord record) {
